Fix float serialization problem

+++ omnigibson/object_states/saturated.py
@@ -226,7 +226,7 @@ class Saturated(RelativeObjectState, BooleanStateMixin):
                 self._limits[k] = v
 
     def serialize(self, state):
-        state_flat = th.tensor([state["n_systems"], state["default_limit"]], dtype=th.float64)
+        state_flat = th.tensor([state["n_systems"], state["default_limit"]], dtype=th.float32)
         if state["n_systems"] > 0:
             system_names = tuple(state.keys())[2:]
             state_flat = th.cat(
@@ -236,7 +236,7 @@ class Saturated(RelativeObjectState, BooleanStateMixin):
                         [
                             th.tensor(
                                 [self.obj.scene.get_system(system_name, force_init=False).uuid, state[system_name]],
-                                dtype=th.float64,
+                                dtype=th.float32,
                             )
                             for system_name in system_names
                         ]
